art2web/requirements.txt
Flask>=2.0
Pillow>=8.0